Provide a LocalIncludeLoader struct for the 'include_loader' option
Instead of passing a 'include_loader' and a separate
'locale_loader_path' option the Mjml.to_html/2 a
Mjml.ParserOptions.LocalIncludeLoader struct is provided to be able to
pass the related configs in one value.

The LocalIncludeLoader has a path field (optional) that can be used to
define a custom include base path.

This approach allows us to provide other include loaders that are
already implemented in mrml in the future without the need to add more
option keys.

The explicite Noop include loader is dropped from the Elixir side,
because it can just be used as the default on the Rust side if no other
valid include loader option was provided.

Invalid include loader options are checked on the Elixir side now
instead of panicking on the Rust side (which also fixes the issue with
the panick message logging in successful test runs).